Added more sfx and improved car physic and collision

+++ DungeonRacer/Source/Ui/HpWidget.cs
@@ -21,9 +21,9 @@ namespace DungeonRacer
 			player.OnModifyHp += HandleModifyHp;
 		}
 
-		private void HandleModifyHp(Player player, int delta)
+		private void HandleModifyHp(Player player, float delta)
 		{
-			bar.Percent = player.Hp / ((float) player.MaxHp);
+			bar.Percent = player.Hp / player.MaxHp;
 		}
 	}
 }
